PatrykMilewski suggestions
Took 1 hour 2 minutes
