+++ Manager/src/main/java/io/raspberrywallet/manager/database/WalletEntity.java
@@ -1,6 +1,8 @@
 package io.raspberrywallet.manager.database;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,8 +10,10 @@ import java.util.Objects;
 
 public class WalletEntity {
 
-    @JsonProperty("keyparts")
-	public List<KeyPartEntity> parts = new ArrayList<KeyPartEntity>();
+	@Getter
+	@Setter
+	@JsonProperty("keyparts")
+	private List<KeyPartEntity> parts = new ArrayList<KeyPartEntity>();
 
 	/*
 	* Needed to override this, so `WalletEntity` can be easily compared.
